da bot/requirements.txt
dearpygui>=1.10.0
keyboard>=0.13.5
pyautogui>=0.9.54
opencv-python>=4.8.0.0
numpy>=1.24.0
Pillow>=9.5.0
pytesseract>=0.3.10
pynput>=1.7.6
flet>=0.24.0
